chore(runtime): split regexp grammar helpers
